requirements.txt
tabula_py
pandas~=1.1.5
requests~=2.31.0
PyMuPDF
fitz~=0.0.1.dev2
numpy
openpyxl